split Jira tracker functionality into a separate file
