+++ apps/trackers/tests/test_service.py
@@ -8,7 +8,7 @@ BTS instances for generating new cassettes.
 
 import pytest
 
-from apps.trackers.bts_tracker import JiraTracker
+from apps.trackers.jira import JiraTracker
 from apps.trackers.models import JiraProjectFields
 from apps.trackers.service import JiraTrackerQuerier
 from apps.trackers.tests.test_bts_tracker import validate_minimum_key_value
